backend/requirements.txt
fastapi
uvicorn[standard]
pillow
supabase
python-dotenv
torch
ultralytics
torchvision
opencv-python-headless
pydantic
python-multipart
email-validator
httpx
PyJWT
